Move Handle<T> to common
